==> 01_Extraccion_WebScraping.ipynb <==
import pandas as pd
from string import ascii_uppercase as alfabeto
import pickle
import io
import requests
#Configuracion que utilizare para evitar bloqueo de seguridad de wikipedia
url = "https://en.wikipedia.org/wiki/2026_FIFA_World_Cup"
headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
          }
response = requests.get(url, headers=headers)

#Extracción de las tablas HTML del mundial 2026
tablas_mundial=pd.read_html(io.StringIO(response.text))
grupos={}

#Limpieza y estructuración, iterando de 7 en 7 que son las tablas de los grupos (A-L)
for letra, i in zip(alfabeto, range(25,105,7)):
    df=tablas_mundial[i]
    df.rename(columns={df.columns[1]:"Equipo"},inplace=True) #Renombro la columna
    df.pop("Qualification") #Eliminno columna inecesaria para mi proyecto
    grupos[f"Grupo {letra}"]=df


#Guardo los datos para después
with open("Grupos2026","wb") as output:
    pickle.dump(grupos,output)
