==> 02_WebScraping_Historico_BS4.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd

#Lista de los años de todos los mundiales
años=[1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
       1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
       2018, 2022]

#Función para obtener todos los partidos que se jugaron en cada año
def obtener_partidos(año):
    web=f"https://en.wikipedia.org/wiki/{año}_FIFA_World_Cup"
    headers = {
        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
    }
    respuesta=requests.get(web,headers=headers)
    contenido=respuesta.text

    
    sopa=BeautifulSoup(contenido,"lxml")
    partidos=sopa.find_all("div",class_="footballbox")

    local=[]
    marcador=[]
    visitante=[]
    for partido in partidos:
        local.append(partido.find("th",class_="fhome").get_text())
        marcador.append(partido.find("th",class_="fscore").get_text())
        visitante.append(partido.find("th",class_="faway").get_text())

    dict_futbol={"Local" : local, "Marcador" : marcador, "Visitante" : visitante}
    dict_futbol["Año"]=año
    df=pd.DataFrame(dict_futbol)
    return df


lista_dataframes = [] # 1. Creamos una lista vacía para guardar los df de los partidos

#Ciclo para automatizar la extracción de los partidos por año
for año in años:
    print(f"Descargando año {año}...") 
    df_temporal = obtener_partidos(año)
    lista_dataframes.append(df_temporal) # 2. Guardamos cada DF en la lista

#Juntamos todos los partidos en un df total
df_total = pd.concat(lista_dataframes, ignore_index=True)

# 4. Guardar a CSV
df_total.to_csv("todos_los_mundiales.csv", index=False)




#Extraigo las tablas de partidos del mundial 2026 que se usara para la predicción
df_fixtures=obtener_partidos(2026)
df_fixtures.to_csv("fixtures.csv",index=False)
