==> 03_DataWrangling_Limpieza.ipynb <==
import pandas as pd

partidos=pd.read_csv("todos_los_mundiales.csv")
fixtures=pd.read_csv("fixtures.csv")

fixtures["Local"]=fixtures["Local"].str.strip() #con .strip() Limpiamos espacios en blanco,saltos de linea etc
fixtures["Visitante"]=fixtures["Visitante"].str.strip()

partidos["Local"]=partidos["Local"].str.strip()
partidos["Visitante"]=partidos["Visitante"].str.strip()

partidos.dropna(inplace=True)#Elimina datos faltantes (NaN)
partidos[partidos["Marcador"].isnull()]
partidos.drop_duplicates(inplace=True)#Eliminamos duplicados
index_eliminar=partidos[partidos["Local"].str.contains("Sweden") & partidos["Visitante"].str.contains("Austria")].index
partidos.drop(index=index_eliminar,inplace=True)

partidos[partidos["Marcador"].str.contains("[^\d-]")]
partidos["Marcador"]=partidos["Marcador"].str.replace("[^\d–]","",regex=True)
partidos[partidos["Marcador"].str.contains("[^\d-]")]
partidos

partidos[["Goles Local","Goles Visitante"]]=partidos["Marcador"].str.split("[–]",expand=True)
partidos.drop("Marcador",axis=1,inplace=True)

partidos=partidos.astype({"Goles Local":int,"Goles Visitante":int})

partidos.dtypes

partidos["Goles Totales"]=partidos["Goles Local"]+partidos["Goles Visitante"]

partidos.to_csv("Mundiales limpio.csv",index=False)
fixtures.to_csv("fixtures.csv",index=False)

años=[1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
       1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
       2018, 2022]
for año in años:
    print(f"Año: {año} Partidos Jugados: {len(partidos[partidos["Año"]==año])}")
