==> 04_Modelo_Predictivo_Simulacion.ipynb <==
import pandas as pd
import pickle
from scipy.stats import poisson


grupos=pickle.load(open("Grupos2026","rb"))
datos_mundial=pd.read_csv("Mundiales limpio.csv")
fixtures=pd.read_csv("fixtures.csv")

datos_mundial

df_local=datos_mundial[["Local","Goles Local","Goles Visitante"]]
df_visitante=datos_mundial[["Visitante","Goles Local","Goles Visitante"]]
df_visitante

df_local=df_local.rename(columns={"Local":"Equipo","Goles Local":"Goles Anotados","Goles Visitante":"Goles Recibidos"})
df_visitante=df_visitante.rename(columns={"Visitante":"Equipo","Goles Local":"Goles Recibidos","Goles Visitante":"Goles Anotados"})
df_visitante

df_equipos_fuerza=pd.concat([df_local,df_visitante],ignore_index=True).groupby("Equipo").mean()
df_equipos_fuerza

def predecir_puntos (local,visitante):
    if local in df_equipos_fuerza.index and visitante in df_equipos_fuerza.index:
        lamb_local=df_equipos_fuerza.at[local,"Goles Anotados"]*df_equipos_fuerza.at[visitante,"Goles Recibidos"]
        lamb_visitante=df_equipos_fuerza.at[visitante,"Goles Anotados"]*df_equipos_fuerza.at[local,"Goles Recibidos"]
        prob_local,prob_visitante,prob_empate=0,0,0
        for x in range(0,11):
            for y in range(0,11):
                p=poisson.pmf(x,lamb_local)*poisson.pmf(y,lamb_visitante)
                if x==y:
                    prob_empate+=p
                elif x>y:
                    prob_local+=p
                else:
                    prob_visitante+=p
        puntos_local=3*prob_local+prob_empate
        puntos_visitante=3*prob_visitante+prob_empate
        return(puntos_local,puntos_visitante)
    else:
        return(0,0)
        
        


predecir_puntos("Mexico","Argentina")



df_fixture_grupo_48 = fixtures[:72].copy()
df_fixture_ronda_32 = fixtures[72:88].copy()
df_fixture_ronda_16 = fixtures[88:96].copy()
df_fixture_cuartos = fixtures[96:100].copy()
df_fixture_semi = fixtures[100:102].copy()
df_fixture_final = fixtures[102:].copy()


# Limpieza de nombres "sucios" del repechaje
# (Estos son ejemplos, revisa cuáles tienes tú en tu DataFrame)

reemplazos = {
    'UEFA Path A winner': 'Wales',
    'UEFA Path A winner[a]': 'Wales',# O el que tú creas que ganará
    'UEFA Path B winner': 'Ukraine',
    'UEFA Path B winner[a]': 'Ukraine',
    'UEFA Path C winner': 'Georgia',
    'UEFA Path C winner[a]': 'Georgia',
    'UEFA Path D winner': 'Iceland',
    'UEFA Path D winner[a]': 'Iceland',# Quítale también el corchete [a]
    'AFC v CONMEBOL winner': 'Peru',
    'IC Play-off 1 Winner': 'DR Congo',      # O 'Jamaica'
    'IC Path 1 winner': 'DR Congo',
    'IC Path 1 winner[a]': 'DR Congo',
    "Mexico (H)":"Mexico",
    "Canada (H)":"Canada",
    "United States (H)":"United States",
    # Repechaje Intercontinental 2 (Asia/Conmebol/Concacaf)
    'IC Play-off 2 Winner': 'Iraq',          # O 'Bolivia'
    'IC Path 2 winner': 'Iraq',
    'IC Path 2 winner[a]': 'Iraq'
    # Agrega todos los que veas raros
}
for grupo in grupos:
    # 'replace' busca el texto sucio y pone el limpio
    grupos[grupo]['Equipo'] = grupos[grupo]['Equipo'].replace(reemplazos)
    
    # Extra: Quitamos espacios vacíos al inicio o final por si acaso
    grupos[grupo]['Equipo'] = grupos[grupo]['Equipo'].str.strip()

# 1. Reemplazamos en la columna 'home' (Local)
df_fixture_grupo_48['Local'] = df_fixture_grupo_48['Local'].replace(reemplazos)

# 2. Reemplazamos en la columna 'away' (Visitante)
df_fixture_grupo_48['Visitante'] = df_fixture_grupo_48['Visitante'].replace(reemplazos)

# 3. Y MUY IMPORTANTE: Limpia los espacios vacíos que a veces quedan
df_fixture_grupo_48['Local'] = df_fixture_grupo_48['Local'].str.strip()
df_fixture_grupo_48['Visitante'] = df_fixture_grupo_48['Visitante'].str.strip()

for grupo in grupos:
    equipo_en_grupos=grupos[grupo]["Equipo"].values
    df_fixture_group_6 = df_fixture_grupo_48[df_fixture_grupo_48['Local'].isin(equipo_en_grupos)]
    for index, row in df_fixture_group_6.iterrows():
        local,visitante=row["Local"],row["Visitante"]
        puntos_local,puntos_visitante=predecir_puntos(local,visitante)
        grupos[grupo].loc[grupos[grupo]["Equipo"]==local, "Pts"]+=puntos_local
        grupos[grupo].loc[grupos[grupo]["Equipo"]==visitante, "Pts"]+=puntos_visitante
    
    grupos[grupo]=grupos[grupo].sort_values("Pts", ascending=False).reset_index()
    grupos[grupo]=grupos[grupo][["Equipo","Pts"]]
    grupos[grupo]=grupos[grupo].round(0)
grupos       

df_fixture_ronda_32


clasificados = {}
terceros_lugares = []

for grupo, tabla in grupos.items():
    letra_grupo = grupo.split()[-1] 
    
    ganador = tabla.iloc[0]['Equipo']
    segundo = tabla.iloc[1]['Equipo']
    
    clasificados[f'Winner Group {letra_grupo}'] = ganador
    clasificados[f'Runner-up Group {letra_grupo}'] = segundo
    
    info_tercero = tabla.iloc[2]
    info_tercero['Group'] = letra_grupo
    terceros_lugares.append(info_tercero)


df_terceros = pd.DataFrame(terceros_lugares)

df_terceros = df_terceros.sort_values(by='Pts', ascending=False)

mejores_8 = df_terceros.head(8)['Equipo'].tolist()

print(f"Los 8 mejores terceros clasificados son: {mejores_8}")



def obtener_rival(nombre_placeholder):
   
    if nombre_placeholder in clasificados:
        return clasificados[nombre_placeholder]
    

    elif '3rd Group' in nombre_placeholder:
        if len(mejores_8) > 0:
            
            return mejores_8.pop(0) 
        else:
            return "Eliminado" 
            
    
    else:
        return nombre_placeholder


for index, row in df_fixture_ronda_32.iterrows():
    home = row['Local']
    away = row['Visitante']
    
    
    df_fixture_ronda_32.at[index, 'Local'] = obtener_rival(home)
    df_fixture_ronda_32.at[index, 'Visitante'] = obtener_rival(away)


df_fixture_ronda_32["Ganador"]="?"
df_fixture_ronda_32

def ganador(tabla_actualizada):
    for index, row in tabla_actualizada.iterrows():
        local, visitante=row["Local"], row["Visitante"]
        puntos_local, puntos_visitante = predecir_puntos(local, visitante)
        if puntos_local>puntos_visitante:
            ganador=local
        else:
            ganador=visitante
        tabla_actualizada.loc[index, "Ganador"]=ganador    
    return(tabla_actualizada)    

ganador(df_fixture_ronda_32)

df_fixture_ronda_16

def actualizar_tabla(df_ronda_1,df_ronda_2):
    for index, row in df_ronda_1.iterrows():
        ganador=df_ronda_1.loc[index,"Ganador"]
        marcador=df_ronda_1.loc[index,"Marcador"]
        df_ronda_2.replace({f"Winner {marcador}":ganador},inplace=True)
    df_ronda_2["Ganador"]="?"     
    return df_ronda_2

actualizar_tabla(df_fixture_ronda_32,df_fixture_ronda_16)

ganador(df_fixture_ronda_16)

actualizar_tabla(df_fixture_ronda_16,df_fixture_cuartos)

df_fixture_cuartos

ganador(df_fixture_cuartos)

actualizar_tabla(df_fixture_cuartos,df_fixture_semi)

df_fixture_semi

ganador(df_fixture_semi)

actualizar_tabla(df_fixture_semi,df_fixture_final)

ganador(df_fixture_final)
